+++ Sources/GanitEditorUI/SheetCommands.swift
@@ -11,6 +11,7 @@ import AppKit
   func showInterpretation(_ sender: Any?)
   func askAssistant(_ sender: Any?)
   func changeAssistantAnswer(_ sender: Any?)
+  func cancelAssistantRequest(_ sender: Any?)
   func insertReference(_ sender: Any?)
   func nextProblem(_ sender: Any?)
   func previousProblem(_ sender: Any?)